Refactor output message functions and remove versions, pull requests from init output
